Project 1 done

+++ Worley.cpp
@@ -1,6 +1,6 @@
  #include "Worley.h"
 #include <math.h>
- 
+
 
 // private helper functions
 namespace
@@ -28,22 +28,22 @@ namespace
 
 // the function to merge-sort a "cube" of samples into the current best-found
 // list of values.
-    void addSamples(long xi, long maxOrder,
-                    float at, float *F, float(*delta), unsigned long *ID);
-    void addSamples(long xi, long yi, long maxOrder,
-                    float at[2], float *F, float(*delta)[2], unsigned long *ID);
-    void addSamples(long xi, long yi, long zi, long maxOrder,
-                    float at[3], float *F, float(*delta)[3], unsigned long *ID);
+    void addSamples(int xi, int maxOrder,
+                    float at, float *F, float(*delta), unsigned int *ID);
+    void addSamples(int xi, int yi, int maxOrder,
+                    float at[2], float *F, float(*delta)[2], unsigned int *ID);
+    void addSamples(int xi, int yi, int zi, int maxOrder,
+                    float at[3], float *F, float(*delta)[3], unsigned int *ID);
 } // namespace
 
 
 void
-WorleyNoise::noise1D(float at, long maxOrder,
-                     float *F, float(*delta), unsigned long *ID)
+WorleyNoise::noise1D(float at, int maxOrder,
+                     float *F, float(*delta), unsigned int *ID)
 {
     float x2, mx2;
     float newAt;
-    long intAt, i;
+    int intAt, i;
 
     // Initialize the F values to "huge" so they will be replaced by the
     // first real sample tests. Note we'll be storing and comparing the
@@ -93,12 +93,12 @@ WorleyNoise::noise1D(float at, long maxOrder,
 
 
 void
-WorleyNoise::noise2D(float at[2], long maxOrder,
-                     float *F, float(*delta)[2], unsigned long *ID)
+WorleyNoise::noise2D(float at[2], int maxOrder,
+                     float *F, float(*delta)[2], unsigned int *ID)
 {
     float x2, y2, mx2, my2;
     float newAt[2];
-    long intAt[2], i;
+    int intAt[2], i;
 
     // Initialize the F values to "huge" so they will be replaced by the
     // first real sample tests. Note we'll be storing and comparing the
@@ -172,12 +172,12 @@ WorleyNoise::noise2D(float at[2], long maxOrder,
 
 
 void
-WorleyNoise::noise3D(float at[3], long maxOrder,
-                     float *F, float(*delta)[3], unsigned long *ID)
+WorleyNoise::noise3D(float at[3], int maxOrder,
+                     float *F, float(*delta)[3], unsigned int *ID)
 {
     float x2, y2, z2, mx2, my2, mz2;
     float newAt[3];
-    long intAt[3], i;
+    int intAt[3], i;
 
     // Initialize the F values to "huge" so they will be replaced by the
     // first real sample tests. Note we'll be storing and comparing the
@@ -303,11 +303,11 @@ namespace
 {
 
 void
-addSamples(long xi, long maxOrder, float at, float *F, float(*delta), unsigned long *ID)
+addSamples(int xi, int maxOrder, float at, float *F, float(*delta), unsigned int *ID)
 {
     float dx, fx, d2;
-    long count, i, j, index;
-    unsigned long seed, this_id;
+    int count, i, j, index;
+    unsigned int seed, this_id;
 
     // Each cube has a random number seed based on the cube's ID number.
     // The seed might be better if it were a nonlinear hash like Perlin uses
@@ -363,12 +363,12 @@ addSamples(long xi, long maxOrder, float at, float *F, float(*delta), unsigned l
 }
 
 void
-addSamples(long xi, long yi, long maxOrder,
-           float at[2], float *F, float(*delta)[2], unsigned long *ID)
+addSamples(int xi, int yi, int maxOrder,
+           float at[2], float *F, float(*delta)[2], unsigned int *ID)
 {
     float dx, dy, fx, fy, d2;
-    long count, i, j, index;
-    unsigned long seed, this_id;
+    int count, i, j, index;
+    unsigned int seed, this_id;
 
     // Each cube has a random number seed based on the cube's ID number.
     // The seed might be better if it were a nonlinear hash like Perlin uses
@@ -429,12 +429,12 @@ addSamples(long xi, long yi, long maxOrder,
 }
 
 void
-addSamples(long xi, long yi, long zi, long maxOrder,
-           float at[3], float *F, float(*delta)[3], unsigned long *ID)
+addSamples(int xi, int yi, int zi, int maxOrder,
+           float at[3], float *F, float(*delta)[3], unsigned int *ID)
 {
     float dx, dy, dz, fx, fy, fz, d2;
-    long count, i, j, index;
-    unsigned long seed, this_id;
+    int count, i, j, index;
+    unsigned int seed, this_id;
 
     // Each cube has a random number seed based on the cube's ID number.
     // The seed might be better if it were a nonlinear hash like Perlin uses
@@ -479,7 +479,7 @@ addSamples(long xi, long yi, long zi, long maxOrder,
             while (index > 0 && d2 < F[index-1]) index--;
 
             // We insert this new point into slot # <index>
-    
+
             // Bump down more distant information to make room for this new point.
             for (i = maxOrder - 2; i >= index; i--)
             {
